+++ app/src/main/java/com/orange/ods/app/ui/components/textfields/ComponentTextField.kt
@@ -38,8 +38,8 @@ import com.orange.ods.compose.component.chip.OdsChoiceChip
 import com.orange.ods.compose.component.chip.OdsChoiceChipsFlowRow
 import com.orange.ods.compose.component.list.OdsListItem
 import com.orange.ods.compose.component.list.OdsListItemTrailingSwitch
-import com.orange.ods.compose.component.tab.OdsTab
 import com.orange.ods.compose.component.tab.OdsTabRow
+import com.orange.ods.compose.component.tab.OdsTabRowTab
 import com.orange.ods.compose.utilities.composable.Keyboard
 import com.orange.ods.compose.utilities.composable.keyboardAsState
 import kotlinx.coroutines.launch
@@ -88,19 +88,20 @@ private fun TextFieldTextCustomization(textFieldCustomizationState: TextFieldCus
         LocalFocusManager.current.clearFocus()
     }
 
-    OdsTabRow(selectedTabIndex = pagerState.currentPage) {
-        tabs.forEachIndexed { index, customizationTab ->
-            OdsTab(
-                text = stringResource(id = customizationTab.titleRes),
-                selected = pagerState.currentPage == index,
-                onClick = {
-                    scope.launch {
-                        pagerState.animateScrollToPage(index)
-                    }
+    OdsTabRow(
+        selectedTabIndex = pagerState.currentPage,
+        tabs = tabs.mapIndexed { index, customizationTab ->
+            OdsTabRowTab(
+                icon = null,
+                text = stringResource(id = customizationTab.titleRes)
+            ) {
+                scope.launch {
+                    pagerState.animateScrollToPage(index)
                 }
-            )
+            }
         }
-    }
+    )
+
     HorizontalPager(state = pagerState, pageCount = tabs.size) { page ->
         Column {
             tabs[page].Content(textFieldCustomizationState)