+++ lib/src/main/java/com/orange/ods/compose/component/tab/OdsScrollableTabRow.kt
@@ -12,7 +12,6 @@ package com.orange.ods.compose.component.tab
 
 import androidx.annotation.DrawableRes
 import androidx.compose.material.ScrollableTabRow
-import androidx.compose.material.Tab
 import androidx.compose.material.TabRowDefaults
 import androidx.compose.material.TabRowDefaults.tabIndicatorOffset
 import androidx.compose.runtime.Composable
@@ -22,6 +21,7 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.tooling.preview.PreviewParameter
 import com.orange.ods.compose.component.OdsComposable
 import com.orange.ods.compose.component.utilities.Preview
 import com.orange.ods.compose.component.utilities.UiModePreviews
@@ -30,21 +30,21 @@ import com.orange.ods.compose.theme.OdsTheme
 /**
  * <a href="https://system.design.orange.com/0c1af118d/p/513d27-tabs/b/50cb71" class="external" target="_blank">ODS tabs</a>.
  *
- * An OdsScrollableTabRow is a Jetpack Compose [ScrollableTabRow] with the Orange design and theme.
+ * An [OdsScrollableTabRow] is a Jetpack Compose [ScrollableTabRow] with the Orange design and theme.
  * @see ScrollableTabRow documentation
  *
- * @param selectedTabIndex the index of the currently selected tab.
- * @param modifier optional [Modifier] for this TabRow.
- * @param tabs the tabs inside this TabRow. Typically this will be multiple [Tab]s. Each element
- * inside this lambda will be measured and placed evenly across the TabRow, each taking up equal
- * space. Use [OdsTab] to display Orange styled tabs.
+ * @param selectedTabIndex Index of the currently selected tab.
+ * @param tabs List of [OdsTabRowTab] displayed inside the tabs row.
+ * @param modifier [Modifier] applied to the scrollable tabs row.
+ * @param tabIconPosition Controls the position of the icon in the tabs. By default, the icon is displayed above the text.
  */
 @Composable
 @OdsComposable
 fun OdsScrollableTabRow(
     selectedTabIndex: Int,
+    tabs: List<OdsTabRowTab>,
     modifier: Modifier = Modifier,
-    tabs: @Composable () -> Unit
+    tabIconPosition: OdsTabRowTabIcon.Position = OdsTabRowTabIcon.Position.Top
 ) {
     ScrollableTabRow(
         modifier = modifier,
@@ -60,13 +60,13 @@ fun OdsScrollableTabRow(
             }
         },
         divider = {},
-        tabs = tabs
+        tabs = { tabs.forEachIndexed { index, tab -> tab.Content(OdsTabRowTab.ExtraParameters(selected = index == selectedTabIndex, iconPosition = tabIconPosition)) } }
     )
 }
 
 @UiModePreviews.Default
 @Composable
-private fun PreviewOdsScrollableTabRow() = Preview {
+private fun PreviewOdsScrollableTabRow(@PreviewParameter(OdsTabRowPreviewParameterProvider::class) parameter: OdsTabPreviewParameter) = Preview {
     data class Tab(@DrawableRes val iconResId: Int, val text: String)
 
     val tabs = listOf(
@@ -77,14 +77,17 @@ private fun PreviewOdsScrollableTabRow() = Preview {
     )
 
     var selectedTabIndex by remember { mutableStateOf(0) }
-    OdsScrollableTabRow(selectedTabIndex = selectedTabIndex) {
-        tabs.forEachIndexed { index, tab ->
-            OdsTab(
-                selected = selectedTabIndex == index,
-                onClick = { selectedTabIndex = index },
-                text = tab.text,
-                icon = painterResource(id = tab.iconResId)
-            )
-        }
+    with(parameter) {
+        OdsScrollableTabRow(
+            selectedTabIndex = selectedTabIndex,
+            tabs = tabs.mapIndexed { index, tab ->
+                OdsTabRowTab(
+                    icon = if (hasIcon) OdsTabRowTabIcon(painterResource(id = tab.iconResId), "") else null,
+                    text = if (hasText) tab.text else null,
+                    enabled = enabled,
+                    onClick = { selectedTabIndex = index }
+                )
+            }
+        )
     }
 }